DeleteAccount delete function works

+++ src/views/DeleteAccount.js
@@ -18,40 +18,22 @@ export default class DeleteAccount extends React.Component {
       .then(data => this.setState({ person: data }));
   }
 
-  firstDelete() {
-    fetch(
-      `https://bookstry20191122022423.azurewebsites.net/api/person/firstDel/` +
-        localStorage.getItem("personId"),
-      {
-        method: "DELETE",
-        headers: {
-          Accept: "application/json",
-          "Content-type": "application/json"
+  deleteAccount() {
+    if (window.confirm("Are you sure you want to delete your account?")) {
+      fetch(
+        `https://bookstry20191122022423.azurewebsites.net/api/person/delAccount/` + localStorage.getItem("personId"),
+        {
+          method: "DELETE",
+          header: {
+            Accept: "application/json",
+            "Content-Type": "application/json"
+          }
         }
-      }
-    );
-  }
-
-  secondDelete() {
-    fetch(
-      `https://bookstry20191122022423.azurewebsites.net/api/person/secondDel/` + localStorage.getItem("personId"),
-      {
-        method: "DELETE",
-        headers: {
-          Accept: "application/json",
-          "Content-type": "application/json"
-        }
-      }
-    );
-  }
-
-  DeleteAccounts() {
-    var firstDel = this.firstDelete();
-    if (firstDel === true) {
-      this.secondDelete();
-    }    
-    else if (firstDel !== true) 
-    alert("Something when wrong")
+      );
+    } else {
+      console.log("cancelled");
+    }
+    window.location.href = "/home-page";
   }
 
   close() {
@@ -59,16 +41,16 @@ export default class DeleteAccount extends React.Component {
   }
 
   render() {
-     var { person } = this.state;
+
     return (
       <Card className="card-register">
         <p>
           By clicking "Delete account permanately", you aggree to the following:
           Your account and all related orders and books will be deleted. Keep in
           mind that if you do so, you will no longer have access to your account
-          and purchases. Your account cannot be restored. {person.personId}
+          and purchases. Your account cannot be restored.
         </p>
-        <Button block color="danger" type="sumbit" onClick={this.DeleteAccounts.bind(this)}>
+        <Button block color="danger" type="sumbit" onClick={this.deleteAccount}>
           Delete account permanately
         </Button>
         <Button block color="secondary" type="submit" onClick={this.close}>
